Extracted CoverageSetting interface
In order to be able to create mocked version for tests.
