+++ src/test/java/com/godaddy/sonar/ruby/simplecovrcov/DefaultCoverageSettingsTest.java
@@ -12,44 +12,44 @@ import static org.junit.Assert.*;
 /**
  * Created by sergio on 3/27/17.
  */
-public class CoverageSettingsTest {
+public class DefaultCoverageSettingsTest {
     private Settings settings;
-    private CoverageSettings coverageSettings;
+    private DefaultCoverageSettings defaultCoverageSettings;
 
     @Before
     public void setUp() throws Exception {
         this.settings = new Settings();
-        this.coverageSettings = new CoverageSettings(settings);
+        this.defaultCoverageSettings = new DefaultCoverageSettings(settings);
     }
 
     @Test
     public void testProcessAllSuitesWhenPropertyIsEmpty() {
         settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, "");
-        assertTrue(coverageSettings.processAllSuites());
+        assertTrue(defaultCoverageSettings.processAllSuites());
     }
 
     @Test
     public void testProcessAllSuitesWhenPropertyContainsAllFlag() {
-        settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, CoverageSettings.ALL_SUITES_FLAG);
-        assertTrue(coverageSettings.processAllSuites());
+        settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, DefaultCoverageSettings.ALL_SUITES_FLAG);
+        assertTrue(defaultCoverageSettings.processAllSuites());
     }
 
     @Test
     public void testProcessAllSuitesWhenPropertyContainsSomeOtherValue() {
         settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, "my_suite");
-        assertFalse(coverageSettings.processAllSuites());
+        assertFalse(defaultCoverageSettings.processAllSuites());
     }
 
     @Test
     public void testConfiguredSuitesNamesWhenAllSuitesSelected() {
-        settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, CoverageSettings.ALL_SUITES_FLAG);
-        assertTrue(coverageSettings.configuredSuitesNames().isEmpty());
+        settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, DefaultCoverageSettings.ALL_SUITES_FLAG);
+        assertTrue(defaultCoverageSettings.configuredSuitesNames().isEmpty());
     }
 
     @Test
     public void testConfiguredSuitesNamesWhenPropertyContainsSomeOtherValue() {
         settings.setProperty(RubyPlugin.COVERAGE_TEST_SUITES_PROPERTY, "MiniSpec,RSpec");
-        List<String> suites = coverageSettings.configuredSuitesNames();
+        List<String> suites = defaultCoverageSettings.configuredSuitesNames();
         assertFalse(suites.isEmpty());
         assertTrue(suites.contains("MiniSpec"));
         assertTrue(suites.contains("RSpec"));
