+++ src/test/resources/test-data/simple_cov_results_multiple_suites.json
@@ -1,0 +1,66 @@
+{
+  "MiniTest": {
+    "coverage": {
+      "/project/source/subdir/file.rb": [
+        1,
+        1,
+        null,
+        1,
+        null,
+        1,
+        null,
+        1,
+        1,
+        1,
+        null,
+        null,
+        1,
+        1,
+        1,
+        null,
+        null,
+        1,
+        1,
+        null,
+        null,
+        1,
+        0,
+        null,
+        null
+      ]
+    },
+    "timestamp": 1373322213
+  },
+  "RSpec": {
+    "coverage": {
+      "/project/source/subdir/file.rb": [
+        1,
+        1,
+        1,
+        1,
+        null,
+        1,
+        null,
+        1,
+        1,
+        1,
+        null,
+        null,
+        1,
+        1,
+        1,
+        null,
+        null,
+        1,
+        1,
+        null,
+        null,
+        1,
+        0,
+        1,
+        null
+      ]
+    },
+    "timestamp": 1373322213
+  }
+}
